requirements.txt
channels==3.0.2
Django==3.1.3
pyguacamole>=0.10
PyYAML>=5.3